5425. move connect to LegendBuilder
